requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
dlt[postgres,sqlalchemy]>=1.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0
pyarrow>=14.0.0
cryptography>=42.0.0
